Revert "Merge branch 'main' into duplicated-benchmarking-results"
This reverts commit 47d0558b0f047aff3ae933afb4bc8a0be551dafb, reversing
changes made to e3ae3bbbda715f1ce0ab353e4ab28587304b98ce.
