+++ .all-contributorsrc
@@ -3596,15 +3596,6 @@
         "code",
         "bug"
       ]
-    },
-    {
-      "login": "Virgile-Foussereau",
-      "name": "Virgile Foussereau",
-      "avatar_url": "https://avatars.githubusercontent.com/u/99748194?v=4",
-      "profile": "https://github.com/Virgile-Foussereau",
-      "contributions": [
-        "code"
-      ]
     }
   ]
 }
