+++ .github/workflows/test_base.yml
@@ -10,7 +10,7 @@ jobs:
     outputs:
       base: ${{ steps.filter.outputs.base }}
     steps:
-      - uses: actions/checkout@v5
+      - uses: actions/checkout@v4
       - uses: dorny/paths-filter@v3
         id: filter
         with:
@@ -32,14 +32,14 @@ jobs:
           - "3.12"
           - "3.13"
         operating-system:
-          - macos-latest
+          - macos-13
           - ubuntu-latest
           - ubuntu-22.04-arm
           - windows-latest
     runs-on: ${{ matrix.operating-system }}
     steps:
       - name: checkout pull request branch
-        uses: actions/checkout@v5
+        uses: actions/checkout@v4
 
       - name: run tests on python ${{ matrix.python-version }}
         uses: ./.github/actions/test-base
