+++ .github/workflows/test_datasets.yml
@@ -17,14 +17,14 @@ jobs:
           - "3.12"
           - "3.13"
         operating-system:
-          - macos-latest
+          - macos-13
           - ubuntu-latest
           - ubuntu-22.04-arm
           - windows-latest
     runs-on: ${{ matrix.operating-system }}
     steps:
       - name: repository checkout step
-        uses: actions/checkout@v5
+        uses: actions/checkout@v4
       - name: update tracking reference step
         run: git remote set-branches origin main
         shell: bash
@@ -54,14 +54,14 @@ jobs:
           - "3.12"
           - "3.13"
         operating-system:
-          - macos-latest
+          - macos-13
           - ubuntu-latest
           - ubuntu-22.04-arm
           - windows-latest
     runs-on: ${{ matrix.operating-system }}
     steps:
       - name: repository checkout step
-        uses: actions/checkout@v5
+        uses: actions/checkout@v4
       - name: update tracking reference step
         run: git remote set-branches origin main
         shell: bash
