+++ .github/workflows/test_module.yml
@@ -10,7 +10,7 @@ jobs:
     outputs:
       module_changes: ${{ steps.filter.outputs.changes }}
     steps:
-      - uses: actions/checkout@v5
+      - uses: actions/checkout@v4
       - uses: dorny/paths-filter@v3
         id: filter
         with:
@@ -65,7 +65,7 @@ jobs:
           - "3.12"
           - "3.13"
         operating-system:
-          - macos-latest
+          - macos-13
           - ubuntu-latest
           - ubuntu-22.04-arm
           - windows-latest
@@ -73,7 +73,7 @@ jobs:
     runs-on: ${{ matrix.operating-system }}
     steps:
       - name: checkout pull request branch
-        uses: actions/checkout@v5
+        uses: actions/checkout@v4
 
       - name: ${{ matrix.sktime-component }}, ${{ matrix.operating-system }}, ${{ matrix.python-version }}
         uses: ./.github/actions/test-component
