+++ .github/workflows/update_contributors.yml
@@ -11,7 +11,7 @@ jobs:
   generate-markdown-and-commit:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v5
+      - uses: actions/checkout@v4
 
       - name: Set up Node
         uses: actions/setup-node@v4
