+++ docs/source/api_reference/forecasting.rst
@@ -655,13 +655,6 @@ Ensembles and stacking
     AutoEnsembleForecaster
     StackingForecaster
 
-.. currentmodule:: sktime.forecasting.residual_booster
-
-.. autosummary::
-    :toctree: auto_generated/
-    :template: class.rst
-
-    ResidualBoostingForecaster
 
 .. currentmodule:: sktime.forecasting.autots
