+++ sktime/classification/tests/_classification_test_reproduction.py
@@ -19,6 +19,7 @@ from sktime.classification.feature_based import (
     Catch22Classifier,
     MatrixProfileClassifier,
     RandomIntervalClassifier,
+    SignatureClassifier,
     SummaryClassifier,
 )
 from sktime.classification.hybrid import HIVECOTEV1, HIVECOTEV2
@@ -274,6 +275,22 @@ if __name__ == "__main__":
                 ),
                 estimator=RandomForestClassifier(n_estimators=10),
                 random_state=0,
+            )
+        ),
+    )
+    _print_array(
+        "SignatureClassifier - UnitTest",
+        _reproduce_classification_unit_test(
+            SignatureClassifier(
+                estimator=RandomForestClassifier(n_estimators=10), random_state=0
+            )
+        ),
+    )
+    _print_array(
+        "SignatureClassifier - BasicMotions",
+        _reproduce_classification_basic_motions(
+            SignatureClassifier(
+                estimator=RandomForestClassifier(n_estimators=10), random_state=0
             )
         ),
     )
