+++ sktime/tests/_config.py
@@ -50,6 +50,7 @@ EXCLUDE_ESTIMATORS = [
     # DL classifier suspected to cause hangs and memouts, see #4610
     "EditDist",
     "CNNClassifier",
+    "InceptionTimeClassifier",
     "LSTMFCNClassifier",
     "MLPClassifier",
     "MLPRegressor",
@@ -107,6 +108,11 @@ EXCLUDED_TESTS = {
     "StackingForecaster": ["test_predict_time_index_with_X"],
     # known side effects on multivariate arguments, #2072
     "WindowSummarizer": ["test_methods_have_no_side_effects"],
+    # tagged in issue #2490
+    "SignatureClassifier": [
+        "test_classifier_on_unit_test_data",
+        "test_classifier_on_basic_motions",
+    ],
     # pickling problem with local method see #2490
     "ProximityStump": [
         "test_persistence_via_pickle",
@@ -151,6 +157,9 @@ EXCLUDED_TESTS = {
         "test_fit_idempotent",
     ],
     "MLPRegressor": [
+        "test_fit_idempotent",
+    ],
+    "InceptionTimeClassifier": [
         "test_fit_idempotent",
     ],
     "SimpleRNNClassifier": [
@@ -350,6 +359,8 @@ EXCLUDED_TESTS_BY_TEST = {
         "SeededBinarySegmentation",
         "ShapeletTransform",
         "ShapeletTransformClassifier",
+        "SignatureClassifier",
+        "SignatureTransformer",
         "SlidingWindowSegmenter",
         "SlopeTransformer",
         "StackingForecaster",
