+++ sktime/transformations/panel/signature_based/_compute.py
@@ -34,6 +34,7 @@ class _WindowSignatureTransform(BaseTransformer):
         "y_inner_mtype": "None",  # which mtypes do _fit/_predict support for X?
         "fit_is_empty": True,
         "python_dependencies": "esig",
+        "python_version": "<3.10",
     }
 
     def __init__(
